Delete tools.test.h
